Refactor

+++ src/DonateTo.ApplicationCore/Entities/Question.cs
@@ -9,8 +9,8 @@ namespace DonateTo.ApplicationCore.Entities
         public string Placeholder { get; set; }
         public string DefaultValue { get; set; }
         public decimal Weight { get; set; }
-        public int Min { get; set; }
-        public int Max { get; set; }
+        public decimal Min { get; set; }
+        public decimal Max { get; set; }
         public long ControlTypeId { get; set; }
         public ControlType ControlType { get; set; }
         public IEnumerable<QuestionOption> Options { get; set; }
