+++ src/DonateTo.ApplicationCore/Entities/QuestionOption.cs
@@ -5,8 +5,8 @@
         public string Label { get; set; }
         public string Value { get; set; }
         public decimal Weight { get; set; }
-        public int MinimumRelative { get; set; }
-        public int MaximumRelative { get; set; }
+        public decimal MinimumRelative { get; set; }
+        public decimal MaximumRelative { get; set; }
         public long QuestionId { get; set; }
         public Question Question { get; set; }
     }
